# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Smart ERP API
  version: 1.1.0
  summary: Machine form of docs/spec/04-api-contracts.md
  description: |
    Seeded from `docs/spec/04-api-contracts.md` version 1.1.0. That file is the human form and
    wins on conflict until this document is regenerated. Additive changes bump the minor version;
    breaking changes bump the major version and carry a migration note.

    1.1.0 (additive): `GET /me/sessions` and `DELETE /me/sessions/{id}` plus schema `UserSession`
    (R1.15, A9). Period soft-close, hard-close, audit-adjustment open, and the holiday calendar.
    No breaking change from 1.0.0.

    Transport rules that apply to every operation:

    - JSON over HTTPS. All timestamps are UTC ISO 8601. Money is `{ amount, currency }` with the
      amount as a decimal string.
    - Every success response is `{ data, meta }`. Every failure, including 429 and gateway errors,
      is `{ error: { code, message, details, request_id } }`.
    - Every mutating request accepts `Idempotency-Key` (UUID); the server replays a stored response
      for 24 hours; a different body with the same key is `409 CONFLICT`.
    - Every mutable resource carries `state_version`; mutating requests send `If-Match`; a mismatch
      is `409 CONFLICT` with the current state in `error.details`.
    - `Accept-Language: en | ar` affects messages and print formats, never data.
  contact:
    name: Smart ERP contract owner
  license:
    name: Proprietary
...
servers:
...
security:
  - bearer: []
    dpop: []
...
paths:
...
  /periods/{id}/soft-close:
    post:
      operationId: softClosePeriod
      tags: [periods]
      summary: Soft-close a fiscal period
      description: |
        Accountant only (R4.6, C3, C33). Blocks further postings into the period. Idempotent
        when the period is already soft-closed and `If-Match` is the current `state_version`.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchRequired"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: Period soft-closed
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/Period"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /periods/{id}/hard-close:
    post:
      operationId: hardClosePeriod
      tags: [periods]
      summary: Hard-close a fiscal period
      description: |
        Stakeholder only, and only with an approval id that the approvals module has granted
        (R4.6, C3). Nothing posts to a hard-closed period; the posting guard returns
        `PERIOD_CLOSED`.
      parameters:
        - $ref: "#/components/parameters/PeriodId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchRequired"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeriodApproval"
      responses:
        "200":
          description: Period hard-closed
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/Period"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /periods/{year}/audit-adjustment/open:
    post:
      operationId: openAuditAdjustmentPeriod
      tags: [periods]
      summary: Open the audit-adjustment period for a fiscal year
      description: |
        Stakeholder only, with approval (R4.8). One audit-adjustment period per fiscal year.
        Opening an already-open period with the same `If-Match` replays the current period.
      parameters:
        - $ref: "#/components/parameters/FiscalYear"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchRequired"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeriodApproval"
      responses:
        "200":
          description: Audit-adjustment period opened
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/Period"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /journeys:
    get:
      operationId: listJourneys
      tags: [journeys]
      summary: List journey definitions grouped by persona
      description: |
        Server-enforced persona filter (R15.5). `persona` is intersected with the personas the
        server has evaluated for the caller. Requesting a persona the caller does not hold is
        `403 PERMISSION_DENIED`. Omitting `persona` returns every group the caller holds.
        Client-supplied roles or personas never widen this list.
      parameters:
        - name: persona
          in: query
          required: false
          description: Persona group to return. Ignored as a grant; used only as a filter.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: Definitions grouped by persona
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/JourneyList"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /journeys/{slug}/instances:
    post:
      operationId: createJourneyInstance
      tags: [journeys]
      summary: Start a journey instance
      description: |
        Creates a persisted instance at the definition's first step and returns
        `{ instance_id, step, state_version }`. `state_version` is 1. The caller must hold a
        persona the definition allows; that check uses server-evaluated personas.
      parameters:
        - $ref: "#/components/parameters/JourneySlug"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "201":
          description: Instance created
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/JourneyInstanceCreated"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /journeys/instances/{id}:
    get:
      operationId: getJourneyInstance
      tags: [journeys]
      summary: Read resumable journey instance state
      description: |
        Returns the persisted instance. This is the resume point after a process restart or after
        days. There is no in-memory journey state.
      parameters:
        - $ref: "#/components/parameters/JourneyInstanceId"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: Instance
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/JourneyInstance"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /journeys/instances/{id}/step:
    post:
      operationId: submitJourneyStep
      tags: [journeys]
      summary: Submit the current journey step
      description: |
        Body `{ step_id, input }`. Requires `Idempotency-Key` and `If-Match` carrying the instance
        `state_version`. A mismatch is `409 CONFLICT` with the current instance in
        `error.details.current`.

        HTTP 200 returns the step result inside the success envelope even when the step did not
        complete: `{ ok, code?, message?, data?, problems[], next_step? }`. `PENDING` means an
        await step is still waiting and the run has not advanced. `REJECTED` means the awaited
        decision went against the run and the run has stopped; later steps do not execute.

        `input` may contain business fields declared by `input_schema`. Claims about permissions,
        roles, personas, or workflow state are ignored and cannot change server-evaluated access
        or workflow state (E13). Permissions are re-evaluated on every step.
      parameters:
        - $ref: "#/components/parameters/JourneyInstanceId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: If-Match
          in: header
          required: true
          description: Current `state_version` of the instance.
          schema:
            type: string
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JourneyStepSubmission"
      responses:
        "200":
          description: Step result (completion, pending, or a business refusal)
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/JourneyStepResult"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ERP access token issued by POST /auth/token. Expires in 15 minutes or less.
    dpop:
      type: apiKey
      in: header
      name: DPoP
      description: |
        DPoP proof JWT signed by the enrolled device key (RFC 9449). Required on every mobile
        request alongside the bearer token; the console uses a session cookie with a CSRF token.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        UUID chosen by the client. The server stores the response for 24 hours and replays it for a
        repeated key with the same body; a different body with the same key is `409 CONFLICT`.
      schema:
        type: string
        format: uuid
...
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: "`en` or `ar`. Affects messages and print formats, never data."
      schema:
        type: string
        enum: [en, ar]
...
    PeriodId:
      name: id
      in: path
      required: true
      description: Fiscal period id
      schema:
        type: string
        format: uuid
    FiscalYear:
      name: year
      in: path
      required: true
      description: Fiscal year, four digits
      schema:
        type: integer
        minimum: 2000
        maximum: 2200
    IfMatchRequired:
      name: If-Match
      in: header
      required: true
      description: |
        The `state_version` the client last saw. Required on period and holiday-calendar
        writes. A mismatch is `409 CONFLICT` with the current state in `error.details.current`.
        The first holiday-calendar write sends `0`.
      schema:
        type: string
    JourneySlug:
      name: slug
      in: path
      required: true
      description: Journey definition slug
      schema:
        type: string
        pattern: "^[a-z][a-z0-9-]*$"
    JourneyInstanceId:
      name: id
      in: path
      required: true
      description: Journey instance id
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Validation failed (`VALIDATION_ERROR`) or a required header is missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: "`AUTH_REQUIRED`, `TOKEN_EXPIRED`, `TOKEN_REVOKED` or `DEVICE_MISMATCH`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: "`PERMISSION_DENIED`, `STEP_UP_REQUIRED` or `SOD_VIOLATION`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: "`NOT_FOUND`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: |
        `CONFLICT` (stale `If-Match`, current state in `error.details.current`, or a reused
        `Idempotency-Key` with a different body), `IMMUTABLE` or `PERIOD_CLOSED`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: "`RATE_LIMITED`, in the standard envelope (A7, A17)"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the client may retry
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: "`INTERNAL_ERROR`; the message never leaks internals, `request_id` correlates logs"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unavailable:
      description: "`INTERNAL_ERROR` or `MISSING_CONFIG` while the service is not ready; also returned by the gateway in the same envelope"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    Meta:
      type: object
      description: Present on every success response.
      required: [as_of, request_id]
      properties:
        as_of:
          type: string
          format: date-time
          description: Server time the response was computed, UTC
        request_id:
          type: string
          description: Correlates the response with logs and audit rows
      additionalProperties: true
...
    ErrorCode:
      type: string
      description: The closed set of error codes from docs/spec/04-api-contracts.md.
      enum:
        - AUTH_REQUIRED
        - TOKEN_EXPIRED
        - TOKEN_REVOKED
        - DEVICE_MISMATCH
        - STEP_UP_REQUIRED
        - PERMISSION_DENIED
        - VALIDATION_ERROR
        - MISSING_CONFIG
        - NOT_FOUND
        - CONFLICT
        - ALREADY_DECIDED
        - IMMUTABLE
        - PERIOD_CLOSED
        - SOD_VIOLATION
        - CREDIT_HOLD
        - PRICE_FLOOR_HOLD
        - NEGATIVE_STOCK
        - RATE_LIMITED
        - INTERNAL_ERROR
    Error:
      type: object
      required: [code, message, details, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human readable, localised by Accept-Language
        details:
          type: object
          description: Code-specific detail, e.g. `{ field }` for VALIDATION_ERROR or `{ current }` for CONFLICT
          additionalProperties: true
        request_id:
          type: string
    ErrorEnvelope:
      type: object
      description: Returned on every failure path including 429 and gateway errors (A17).
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    StateVersion:
      type: integer
      minimum: 0
      description: Optimistic concurrency version; bumped on every transition
...
    PeriodStatus:
      type: string
      enum: [open, soft_closed, hard_closed]
    PeriodKind:
      type: string
      enum: [month, audit_adjustment]
    PeriodApproval:
      type: object
      description: Approval id granted by the approvals module before a Stakeholder close or open.
      required: [approval_id]
      additionalProperties: false
      properties:
        approval_id:
          type: string
          minLength: 1
    Period:
      type: object
      required: [id, fiscal_year, kind, status, start_date, end_date, state_version]
      properties:
        id:
          type: string
          format: uuid
        fiscal_year:
          type: integer
        month:
          type: [integer, "null"]
          minimum: 1
          maximum: 12
          description: Null on the audit-adjustment period
        kind:
          $ref: "#/components/schemas/PeriodKind"
        status:
          $ref: "#/components/schemas/PeriodStatus"
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        state_version:
          $ref: "#/components/schemas/StateVersion"
...
    JourneyList:
      type: object
      required: [groups]
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/JourneyGroup"
    JourneyGroup:
      type: object
      required: [persona, definitions]
      properties:
        persona:
          type: string
        definitions:
          type: array
          items:
            $ref: "#/components/schemas/JourneyDefinition"
    JourneyDefinition:
      type: object
      required: [slug, title, group, personas, steps]
      properties:
        slug:
          type: string
          pattern: "^[a-z][a-z0-9-]*$"
        title:
          type: string
          description: Localised by Accept-Language. The stored title is a message key.
        group:
          type: string
        personas:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/JourneyStep"
    JourneyKind:
      type: string
      enum: [form, validate, write_draft, route, await, post, read]
    JourneyStep:
      type: object
      required: [step_id, kind, title, input_schema]
      properties:
        step_id:
          type: string
        kind:
          $ref: "#/components/schemas/JourneyKind"
        title:
          type: string
        input_schema:
          type: object
          description: JSON Schema for the step input. The client does not execute guards.
          additionalProperties: true
        guard:
          type: object
          description: Server-evaluated guard data. Not an expression the client can supply.
          additionalProperties: true
    JourneyInstanceCreated:
      type: object
      required: [instance_id, step, state_version]
      properties:
        instance_id:
          type: string
          format: uuid
        step:
          $ref: "#/components/schemas/JourneyStep"
        state_version:
          $ref: "#/components/schemas/StateVersion"
    JourneyInstanceStatus:
      type: string
      enum: [running, awaiting, rejected, completed]
    JourneyInstance:
      type: object
      required:
        - instance_id
        - slug
        - persona
        - status
        - current_step
        - state_version
        - server_state
        - updated_at
      properties:
        instance_id:
          type: string
          format: uuid
        slug:
          type: string
        persona:
          type: string
        status:
          $ref: "#/components/schemas/JourneyInstanceStatus"
        current_step:
          $ref: "#/components/schemas/JourneyStep"
        state_version:
          $ref: "#/components/schemas/StateVersion"
        server_state:
          type: object
          description: |
            Written only by the engine. Client input is not copied here until forbidden keys
            (permissions, roles, personas, workflow state) are removed.
          additionalProperties: true
        updated_at:
          type: string
          format: date-time
    JourneyStepSubmission:
      type: object
      required: [step_id, input]
      additionalProperties: false
      properties:
        step_id:
          type: string
        input:
          type: object
          description: |
            Business fields for this step. Keys permissions, roles, personas, workflow_state,
            granted, and status are ignored.
          additionalProperties: true
    JourneyStepCode:
      type: string
      description: |
        Outcome code on a step result. Distinct from the error-envelope ErrorCode set so that
        PENDING and REJECTED stay journey outcomes rather than transport failures.
      enum: [PERMISSION_DENIED, VALIDATION_ERROR, PENDING, REJECTED, CONFLICT]
    JourneyProblem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
          description: Localised by Accept-Language
        field:
          type: string
    JourneyStepResult:
      type: object
      description: |
        Step protocol result. `ok` is false while an await is pending and when the run stops on
        rejection. `problems` is always present and empty when there are none. `next_step` is
        omitted when the run has stopped.
      required: [ok, problems]
      properties:
        ok:
          type: boolean
        code:
          $ref: "#/components/schemas/JourneyStepCode"
        message:
          type: string
        data:
          type: object
          additionalProperties: true
        problems:
          type: array
          items:
            $ref: "#/components/schemas/JourneyProblem"
        next_step:
          $ref: "#/components/schemas/JourneyStep"